check for tests in ci
